Fix Windows/Linux build problems
git-svn-id: http://ludobloom.com/svn/StemLibProjects@373 d291a34c-4c4d-0410-8556-a73de5c6a500

+++ pngimageio/trunk/test_source/testharness/TestHarness_main.c
@@ -154,7 +154,7 @@ int main(int argc, char ** argv) {
 		return EXIT_FAILURE;
 	}
 	if (inputIsRaw && infileLength % (inputWidth * BitmapImage_pixelFormatBytes(inputFormat)) != 0) {
-		fprintf(stderr, "Raw image data length is not divisible by specified pixel width (%lu bytes left at the end of the file when expecting rows of %d bytes)\n", infileLength % (inputWidth * BitmapImage_pixelFormatBytes(inputFormat)), inputWidth * BitmapImage_pixelFormatBytes(inputFormat));
+		fprintf(stderr, "Raw image data length is not divisible by specified pixel width (%d bytes left at the end of the file when expecting rows of %d bytes)\n", (int) (infileLength % (inputWidth * BitmapImage_pixelFormatBytes(inputFormat))), (int) (inputWidth * BitmapImage_pixelFormatBytes(inputFormat)));
 		return EXIT_FAILURE;
 	}
 	if (inputIsRaw && flipInput) {
